Introduce the common function in securities, bank base class
